requirements.txt
# requirements.txt
# Versiones fijadas para asegurar consistencia durante depuración

azure-functions==1.19.0
msal==1.32.3  # Soporta MsalUiRequiredException
requests==2.31.0
azure-identity==1.16.0  # Compatible con msal>=1.24.0
PyJWT[crypto]==2.10.1  # Recomendado para trabajar con tokens OBO